+++ Personal-Ledger/src/main/java/it/jac/ledger/services/TransactionService.java
@@ -35,7 +35,7 @@ public class TransactionService{
 		
 	}
 	
-	public float getBilancioByUsername(String username) {
+	public double getBilancioByUsername(String username) {
 		return this.transDao.getBilancioByUsername(username);
 	}
